refactored rest of modules into their folders
