+++ liquidhaskell.cabal
@@ -147,34 +147,34 @@ Library
                     Language.Haskell.Liquid.Constraint.ToFixpoint,
                     Language.Haskell.Liquid.Measure,
                     Language.Haskell.Liquid.Parse,
-                    Language.Haskell.Liquid.GhcInterface,
+                    Language.Haskell.Liquid.GHC.Interface,
                     Language.Haskell.Liquid.Types.RefType,
-                    Language.Haskell.Liquid.RefSplit,
+                    Language.Haskell.Liquid.Transforms.RefSplit,
                     Language.Haskell.Liquid.UX.Errors,
                     Language.Haskell.Liquid.Types.PredType,
                     Language.Haskell.Liquid.UX.ACSS,
                     Language.Haskell.Liquid.UX.DiffCheck,
-                    Language.Haskell.Liquid.ANFTransform,
+                    Language.Haskell.Liquid.Transforms.ANF,
                     Language.Haskell.Liquid.UX.Annotate,
                     Language.Haskell.Liquid.UX.CTags,
                     Language.Haskell.Liquid.UX.Config,
                     Language.Haskell.Liquid.UX.CmdLine,
                     Language.Haskell.Liquid.UX.Cabal,
-                    Language.Haskell.Liquid.GhcMisc,
-                    Language.Haskell.Liquid.GhcPlay,
+                    Language.Haskell.Liquid.GHC.Misc,
+                    Language.Haskell.Liquid.GHC.Play,
                     Language.Haskell.Liquid.Misc,
-                    Language.Haskell.Liquid.CoreToLogic,
+                    Language.Haskell.Liquid.Transforms.CoreToLogic,
                     Language.Haskell.Liquid.Types.Variance,
                     Language.Haskell.Liquid.Types.Bounds,
                     Language.Haskell.Liquid.Types.Dictionaries,
                     Language.Haskell.Liquid.Constraint.Qualifier,
-                    Language.Haskell.Liquid.TransformRec,
+                    Language.Haskell.Liquid.Transforms.Rec,
                     Language.Haskell.Liquid.UX.Tidy,
                     Language.Haskell.Liquid.Types,
-                    Language.Haskell.Liquid.Simplify,
+                    Language.Haskell.Liquid.Transforms.Simplify,
                     Language.Haskell.Liquid.Types.Literals,
                     Language.Haskell.Liquid.Types.Strata,
-                    Language.Haskell.Liquid.Fresh,
+                    Language.Haskell.Liquid.Constraint.Fresh,
                     Language.Haskell.Liquid.Types.Visitors,
                     Language.Haskell.Liquid.WiredIn,
                     Language.Haskell.Liquid.Types.Names,
