+++ src/Language/Haskell/Liquid/Bare/RTEnv.hs
@@ -15,7 +15,7 @@ import qualified Data.List           as L
 import Language.Fixpoint.Misc (errorstar, fst3)
 import Language.Fixpoint.Types (Expr(..), Pred(..), Symbol)
 
-import Language.Haskell.Liquid.GhcMisc (sourcePosSrcSpan)
+import Language.Haskell.Liquid.GHC.Misc (sourcePosSrcSpan)
 import Language.Haskell.Liquid.Types.RefType (symbolRTyVar)
 import Language.Haskell.Liquid.Types
