+++ src/Language/Haskell/Liquid/Bare/RefToLogic.hs
@@ -14,7 +14,7 @@ import Language.Haskell.Liquid.Bare.Env
 import Language.Fixpoint.Types hiding (Def, R)
 import Language.Fixpoint.Misc  (errorstar)
 import Language.Fixpoint.Types.Names
-import Language.Haskell.Liquid.GhcMisc (dropModuleUnique)
+import Language.Haskell.Liquid.GHC.Misc (dropModuleUnique)
 import qualified Data.HashMap.Strict as M
 
 import Control.Applicative                      ((<$>))
