+++ src/Language/Haskell/Liquid/Bare/Spec.hs
@@ -40,7 +40,7 @@ import Language.Fixpoint.Misc (group, snd3)
 import Language.Fixpoint.Types.Names (dropSym, isPrefixOfSym,  symbolString)
 import Language.Fixpoint.Types (Qualifier(..), symbol)
 import Language.Haskell.Liquid.Types.Dictionaries
-import Language.Haskell.Liquid.GhcMisc ( dropModuleNames, qualifySymbol, takeModuleNames, getSourcePos, showPpr, symbolTyVar)
+import Language.Haskell.Liquid.GHC.Misc ( dropModuleNames, qualifySymbol, takeModuleNames, getSourcePos, showPpr, symbolTyVar)
 import Language.Haskell.Liquid.Misc (addFst3, fourth4, mapFst, concatMapM)
 import Language.Haskell.Liquid.Types.RefType (generalize, rVar, symbolRTyVar)
 import Language.Haskell.Liquid.Types
