+++ src/Language/Haskell/Liquid/Constraint/Generate.hs
@@ -67,7 +67,7 @@ import qualified Language.Haskell.Liquid.UX.CTags      as Tg
 import Language.Fixpoint.SortCheck (pruneUnsortedReft)
 import Language.Fixpoint.Types.Visitor
 import Language.Fixpoint.Types.Names (symbolString)
-import Language.Haskell.Liquid.Fresh
+import Language.Haskell.Liquid.Constraint.Fresh
 
 import qualified Language.Fixpoint.Types            as F
 
@@ -81,13 +81,13 @@ import Language.Haskell.Liquid.Types.Bounds
 import Language.Haskell.Liquid.Types.RefType
 import Language.Haskell.Liquid.Types.Visitors         hiding (freeVars)
 import Language.Haskell.Liquid.Types.PredType         hiding (freeTyVars)
-import Language.Haskell.Liquid.GhcMisc          ( isInternal, collectArguments, tickSrcSpan
-                                                , hasBaseTypeVar, showPpr, isDataConId
-                                                , symbolFastString, stringVar, stringTyVar)
+import Language.Haskell.Liquid.GHC.Misc          ( isInternal, collectArguments, tickSrcSpan
+                                                 , hasBaseTypeVar, showPpr, isDataConId
+                                                 , symbolFastString, stringVar, stringTyVar)
 import Language.Haskell.Liquid.Misc
 import Language.Fixpoint.Misc
 import Language.Haskell.Liquid.Types.Literals
-import Language.Haskell.Liquid.RefSplit
+import Language.Haskell.Liquid.Transforms.RefSplit
 import Control.DeepSeq
 
 import Language.Haskell.Liquid.Constraint.Axioms
