+++ src/Language/Haskell/Liquid/GHC/Interface.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE TupleSections             #-}
 {-# LANGUAGE ScopedTypeVariables       #-}
 
-module Language.Haskell.Liquid.GhcInterface (
+module Language.Haskell.Liquid.GHC.Interface (
 
   -- * extract all information needed for verification
     getGhcInfo
@@ -47,9 +47,9 @@ import Language.Fixpoint.Misc
 
 import Language.Haskell.Liquid.Types
 import Language.Haskell.Liquid.UX.Errors
-import Language.Haskell.Liquid.ANFTransform
+import Language.Haskell.Liquid.Transforms.ANF
 import Language.Haskell.Liquid.Bare
-import Language.Haskell.Liquid.GhcMisc
+import Language.Haskell.Liquid.GHC.Misc
 import Language.Haskell.Liquid.Misc
 import Language.Haskell.Liquid.UX.PrettyPrint
 import Language.Haskell.Liquid.Types.Visitors
