+++ src/Language/Haskell/Liquid/Measure.hs
@@ -37,7 +37,7 @@ import Data.Maybe (fromMaybe)
 
 import Language.Fixpoint.Misc
 import Language.Fixpoint.Types hiding (Def, R)
-import Language.Haskell.Liquid.GhcMisc
+import Language.Haskell.Liquid.GHC.Misc
 import Language.Haskell.Liquid.Types    hiding (GhcInfo(..), GhcSpec (..))
 import Language.Haskell.Liquid.Misc     (mapSnd)
 import Language.Haskell.Liquid.Types.RefType
