+++ src/Language/Haskell/Liquid/Types/Literals.hs
@@ -12,7 +12,7 @@ import qualified TyCon  as TC
 import Language.Haskell.Liquid.Measure
 import Language.Haskell.Liquid.Types
 import Language.Haskell.Liquid.Types.RefType
-import Language.Haskell.Liquid.CoreToLogic (mkLit)
+import Language.Haskell.Liquid.Transforms.CoreToLogic (mkLit)
 
 import qualified Language.Fixpoint.Types as F
