+++ src/Language/Haskell/Liquid/UX/DiffCheck.hs
@@ -45,7 +45,7 @@ import            Language.Fixpoint.Types         (FixResult (..), Located (..))
 import            Language.Fixpoint.Utils.Files
 import            Language.Haskell.Liquid.Types   (SpecType, GhcSpec (..), AnnInfo (..), DataConP (..), Error, TError (..), Output (..))
 import            Language.Haskell.Liquid.Misc    (mkGraph)
-import            Language.Haskell.Liquid.GhcMisc
+import            Language.Haskell.Liquid.GHC.Misc
 import            Language.Haskell.Liquid.Types.Visitors
 import            Language.Haskell.Liquid.UX.Errors   ()
 import            Text.Parsec.Pos                  (sourceName, sourceLine, sourceColumn, SourcePos, newPos)
