+++ src/Language/Haskell/Liquid/WiredIn.hs
@@ -19,7 +19,7 @@ module Language.Haskell.Liquid.WiredIn
 import Language.Haskell.Liquid.Types
 import Language.Haskell.Liquid.Misc (mapSnd)
 import Language.Haskell.Liquid.Types.RefType
-import Language.Haskell.Liquid.GhcMisc
+import Language.Haskell.Liquid.GHC.Misc
 import Language.Haskell.Liquid.Types.Variance
 import Language.Haskell.Liquid.Types.PredType
